반차와 연차 차감 갯수 잘못 적용되던 현상 수정

+++ src/com/konkuk/repository/EmployeeRepository.java
@@ -45,7 +45,7 @@ public class EmployeeRepository extends Repository implements  IEmployeeReposito
         List<String> result = new ArrayList<>();
         result.add(String.valueOf(employee.id));
         result.add(employee.name);
-        result.add(String.valueOf(employee.salary));
+        result.add(String.valueOf((int) employee.salary));
         result.add(String.valueOf(employee.residualDayOff));
         return result;
     }
